Moved sfetch function to its own file, removed deprecated modules, moved boot.rs to main.rs
